+++ spec/unit/hanami/assets/version_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe "Hanami::Assets::VERSION" do
   it "exposes version" do
-    expect(Hanami::Assets::VERSION).to eq("2.1.0.alpha1")
+    expect(Hanami::Assets::VERSION).to eq("2.1.0.beta2")
   end
 end